Move embed modal state vocabulary into the store state types
