+++ frontend/src/metabase/embedding/components/QuestionEmbedWidget/QuestionEmbedWidget.tsx
@@ -3,8 +3,8 @@ import {
   useUpdateCardEnableEmbeddingMutation,
 } from "metabase/api";
 import { EmbedModal } from "metabase/embedding/components/EmbedModal";
-import { STATIC_LEGACY_EMBEDDING_TYPE } from "metabase/embedding/constants";
 import { useSelector } from "metabase/redux";
+import { STATIC_LEGACY_EMBEDDING_TYPE } from "metabase/redux/store/modal";
 import { getMetadata } from "metabase/selectors/metadata";
 import { getCardUiParameters } from "metabase-lib/v1/parameters/utils/cards";
 import type { Card } from "metabase-types/api";
