+++ frontend/src/metabase/embedding/constants.ts
@@ -1,6 +1,5 @@
 import type { EmbeddingType } from "metabase-types/api";
 
-export const STATIC_LEGACY_EMBEDDING_TYPE: EmbeddingType = "static-legacy";
 export const GUEST_EMBED_EMBEDDING_TYPE: EmbeddingType = "guest-embed";
 
 /**
